+++ src/main/java/com/danahub/zipitda/order/dto/CartListResponseDto.java
@@ -1,0 +1,7 @@
+package com.danahub.zipitda.order.dto;
+
+import java.util.List;
+
+public record CartListResponseDto(
+        List<CartResponseDto> items
+) { }
